refactor(dialogue_manager): extract turn.rs
Step 8/9. Move the per-turn hot path (respond_to wrapper,
respond_to_streaming orchestrator, and the seven private
decomposition methods from commit 7) into dialogue_manager/turn.rs.

mod.rs now contains only structs, type aliases, module declarations,
and the test module. The 461-line impl block became turn.rs with the
same content reorganised under its own module-level doc comment
explaining the step sequence and the load-bearing spawn order.

Trims unused top-level imports from mod.rs (Utc, StreamExt,
PedagogicalIntent, Speaker, Turn, Result, GenerationParams,
prompt_builder, KnowledgeBase) — all moved with their usages to
turn.rs / lifecycle.rs / etc. Adds Speaker, Turn, PrimerError,
Result imports to mod tests' use block (previously inherited via
super::*).

mod.rs: 2459 → 666 lines (66% reduction).
turn.rs: new, 480 lines.
